WIP Less per-byte logic in socket-receive
Not sure this is worth the effort, though.  Maybe it is possible to
use static vectors and receive bytes straight to the buffer in
BASE-CHAR and (UNSIGNED-BYTE 8) cases?  Need to check how it is done
in the file streams.

There is also quite a bit of keyword argument processing done.  Maybe
we need a few different low-level functions (one per ELEMENT-TYPE?)
and have the current SOCKET-RECEIVE call into those?  Maybe even have
a compiler macro?

Also not sure how worthwhile is this because we're dealing with I/O,
anyway...