Install SBCL with correct permissions
Use `install' instead of plain `cp'.  This works around restrictive
user's umask if destination file ownership is different than the user
compiling SBCL (i.e., root).  This does not help at all if the
destination files are supposed to honor the user's umask...

To fix this properly we should check if the user installing SBCL is
the same as the user who has compiled it.  If so, we should keep the
old behaviour, and use `install' otherwise.

The change in d93815635 does not help.